backend/requirements.txt
cassandra-driver==3.24.0
fastapi==0.63.0
uvicorn[standard]==0.13.2
